Fix remaining import paths after backend reorganization
- Fixed dataSyncService import in scrapingSchedulerService.js
- Fixed logger, translationService, and enhancedKeywordMatcher imports in categoryAIService.js
- Fixed config and startupInvestmentService imports in server.js
- Backend now starts successfully on port 3001
- All critical import path issues resolved post-reorganization
